alteração do type de alguns atributos de integer para decimal

+++ PatroCars-Crud-Tradicional/db-persistence/tables/montadora.sql
@@ -12,8 +12,8 @@ CREATE TABLE Modelo (
 		id_modelo SERIAL NOT NULL PRIMARY KEY,
 		nome VARCHAR(50) NOT NULL,
 		montadora_id INT NOT NULL,
-		valor_referencia INT NOT NULL,
-		motorizacao INT NOT NULL,
+		valor_referencia DECIMAL(10,2) NOT NULL,
+		motorizacao DECIMAL(2,1) NOT NULL,
 		turbo CHAR(1) NOT NULL,
 		automatico CHAR(1) NOT NULL,
 		CONSTRAINT FK_MODELO FOREIGN KEY(montadora_id) REFERENCES montadora(id_montadora)
@@ -26,7 +26,7 @@ CREATE TABLE Veiculo (
 		cor VARCHAR(50) NOT NULL,
 		ano_fabricacao INT NOT NULL,
 		ano_modelo INT NOT NULL,
-		valor INT NOT NULL,
+		valor DECIMAL(10,2) NOT NULL,
 		placa VARCHAR(50) NOT NULL,
 		vendido CHAR(1) NOT NULL,
 		CONSTRAINT FK_VEICULO FOREIGN KEY(modelo_id) REFERENCES modelo(id_modelo)
